[svn-r21362] Change the integer type for filno in in  H5O_info_t derived type to c_long to match the unsigned long, via standard:
F2003, Note 15.9

  "The C international standard specifies that the representations for
   nonnegative signed integers ar ethe same as the corresponding
   values for signed integers. Because Fortran does not provide direct
   support for unsigned kinds of integers, the ISO_C_BINDING module
   does not make accessible named constants for their kind type
   parameter values. Instead, a user can use the signed kinds of
   integers to interoperate with the unsigned types and all their
   qualified versions as well...." 

Tested: (jam, intel)

+++ fortran/src/H5Off_F03.f90
@@ -85,13 +85,13 @@ MODULE H5O_PROVISIONAL
   
 
   TYPE, BIND(C) :: H5O_info_t
-     INTEGER(HADDR_T) ::  fileno ! File number that object is located in
-                                   ! unsigned long
-     INTEGER(HADDR_T) :: addr    ! Object address in file  
-     INTEGER :: TYPE             ! Basic object type (group, dataset, etc.) 
-                                   ! H5O_type_t          type
-     INTEGER(c_int) :: rc        ! Reference count of object
-                                   ! unsigned rc
+     INTEGER(c_long) ::  fileno ! File number that object is located in
+                                ! unsigned long
+     INTEGER(HADDR_T) :: addr   ! Object address in file  
+     INTEGER(c_int) :: TYPE     ! Basic object type (group, dataset, etc.) 
+                                ! H5O_type_t type which is type enum
+     INTEGER(c_int) :: rc       ! Reference count of object
+                                ! unsigned rc
      INTEGER(c_int) :: atime    ! Access time
      INTEGER(c_int) :: mtime    ! Modification time
      INTEGER(c_int) :: ctime    ! Change time
@@ -241,7 +241,7 @@ CONTAINS
          CHARACTER(LEN=*), INTENT(IN)  :: name
          INTEGER(SIZE_T) , INTENT(IN)  :: namelen
          INTEGER(HID_T)  , INTENT(IN)  :: lapl_id_default
-         TYPE(C_PTR),value                   :: object_info
+         TYPE(C_PTR),value             :: object_info
 
        END FUNCTION h5oget_info_by_name_c
     END INTERFACE
